refactor: remove hardcode node logic
